Fix HP filter : cast to float so 5/10 HP doesn't round down to 0

+++ Source/ListFilter.cs
@@ -288,7 +288,7 @@ namespace List_Everything
 
 		public override bool Applies(Thing thing) =>
 			thing.def.useHitPoints &&
-			range.Includes(thing.HitPoints / thing.MaxHitPoints);
+			range.Includes((float)thing.HitPoints / thing.MaxHitPoints);
 
 		public override bool DrawOption(Rect rect)
 		{
